jetson-tx2: Add sample with xfce & gstreamer
This sample Dockerfile allows running xfce
with keyboard and mouse, using a balenalib
base image.

Balenalib base image allows udev in container
to notify X about connected mice and keyboards.

This sample runs on L4T 28.2 and will be ported
to 32.2 based BalenaOS when it will be available
in production.

Signed-off-by: Alexandru Costache <alexandru@balena.io>